TestTaskDjango/requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
pytest==8.2.0
pytest-django==4.8.0
celery==5.4.0
redis==5.0.4
